# backend/src/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Swagger for a Node.js Api about superheroes",
        "version": "1.0.0",
        "title": "Swagger MyHero",
        "contact": {
            "email": "m.mercedes.perea@gmail.com"
        }
    },
    "host": "backendmyhero.eu-gb.mybluemix.net",
    "basePath": "/api",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/hero/all/MarvelHeroes": {
            "get": {
                "tags": [
                    "hero"
                ],
                "summary": "Bring all MarvelHeroes",
                "description": "Bring all MarvelHeroess in the app",
                "operationId": "allMarvelHeroes",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Hero"
                            }
                        }
                    },
                    "400": {
                        "description": "There is no heros"
                    }
                }
            }
        },
        "/hero/all/DCHeroes": {
            "get": {
                "tags": [
                    "hero"
                ],
                "summary": "Bring all DCHeroes in the app",
                "description": "Bring all DCHeroes in the app",
                "operationId": "allDCHeroes",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Hero"
                            }
                        }
                    },
                    "400": {
                        "description": "There is no heros"
                    }
                }
            }
        },
...
        "/team/bestTeam": {
            "get": {
                "tags": [
                    "team"
                ],
                "summary": "Get team with higher stats",
                "description": "Get team with higher stats",
                "operationId": "bestTeam",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Ids must be sent"
                    }
                }
            }
        },
        "/team/getTeamInfo/{idUsu}": {
            "get": {
                "tags": [
                    "team"
                ],
                "summary": "Get basic info user Team",
                "description": "Get basic info user team",
                "operationId": "GetTeamInfo",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "idUsu",
                        "description": "id of the User",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/team"
                        }
                    },
                    "400": {
                        "description": "Ids must be sent"
                    }
                }
            }
        },
        "/team/createTeam": {
            "post": {
                "tags": [
                    "team"
                ],
                "summary": "Create a team",
                "description": "Create a team",
                "operationId": "createTeam",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "comment a hero",
                        "required": [
                            "idUsu",
                            "teamName"
                        ],
                        "properties": {
                            "idUsu": {
                                "type": "number",
                                "desciption": "id User"
                            },
                            "teamName": {
                                "type": "string",
                                "desciption": "Team name"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/team/addMember/{idTeam}": {
            "put": {
                "tags": [
                    "team"
                ],
                "summary": "Add member in the team",
                "description": "Add member in the team",
                "operationId": "addMember",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idTeam",
                        "in": "path",
                        "description": "id Team",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Member position ej.member_1 and cod Hero",
                        "required": [
                            "member",
                            "codHero"
                        ],
                        "properties": {
                            "member": {
                                "type": "string",
                                "desciption": "member in the team, ex. member_1"
                            },
                            "codHero": {
                                "type": "string",
                                "desciption": "id hero"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "ids must be sent"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/team/deleteMember/{idTeam}": {
            "put": {
                "tags": [
                    "team"
                ],
                "summary": "Delete member in the team",
                "description": "Delete member in the team",
                "operationId": "deleteMember",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idTeam",
                        "in": "path",
                        "description": "id Team",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Member position ej.member_1 and cod Hero",
                        "required": [
                            "member",
                            "codHero"
                        ],
                        "properties": {
                            "member": {
                                "type": "string",
                                "desciption": "member in the team, ex. member_1"
                            },
                            "codHero": {
                                "type": "string",
                                "desciption": "id hero"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "ids must be sent"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/team/{idTeam}": {
            "delete": {
                "tags": [
                    "team"
                ],
                "summary": "Delete team",
                "description": "",
                "operationId": "Delete team",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idTeam",
                        "in": "path",
                        "description": "Team id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "team not found"
                    }
                }
            }
        },
...
        "/user/deleteUser/{idUsu}": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Delete a user",
                "description": "Delete a user",
                "operationId": "DeleteUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idUsu",
                        "in": "path",
                        "description": "id User",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "comment a hero",
                        "required": [
                            "email",
                            "password"
                        ],
                        "properties": {
                            "email": {
                                "type": "string",
                                "desciption": "email"
                            },
                            "password": {
                                "type": "string",
                                "desciption": "old password"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Hero": {
            "type": "object",
            "properties": {
                "idHero": {
                    "type": "integer",
                    "format": "int64"
                },
                "HeroName": {
                    "type": "string",
                    "description": "Hero name"
                },
                "image": {
                    "type": "string",
                    "description": "Hero image"
                },
                "intelligence": {
                    "type": "integer",
                    "format": "int64"
                },
                "strength": {
                    "type": "integer",
                    "format": "int64"
                },
                "speed": {
                    "type": "integer",
                    "format": "int64"
                },
                "durability": {
                    "type": "integer",
                    "format": "int64"
                },
                "power": {
                    "type": "integer",
                    "format": "int64"
                },
                "combat": {
                    "type": "integer",
                    "format": "int64"
                },
                "fullName": {
                    "type": "string",
                    "description": "Hero full name"
                },
                "placeOfBirth": {
                    "type": "string",
                    "description": "place Of Birth"
                },
                "publisher": {
                    "type": "string",
                    "description": "publisher"
                },
                "alignment": {
                    "type": "string",
                    "description": "alignment"
                },
                "firstApperance": {
                    "type": "string",
                    "description": "firstApperance"
                },
                "gender": {
                    "type": "string",
                    "description": "gender"
                },
                "race": {
                    "type": "string",
                    "description": "race"
                },
                "height": {
                    "type": "string",
                    "description": "height"
                },
                "weight": {
                    "type": "string",
                    "description": "weight"
                },
                "eyeColor": {
                    "type": "string",
                    "description": "eyeColor"
                },
                "hairColor": {
                    "type": "string",
                    "description": "hairColor"
                },
                "work": {
                    "type": "string",
                    "description": "work"
                },
                "biography": {
                    "type": "string",
                    "description": "biography"
                },
                "createDate": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
        "team": {
            "type": "object",
            "properties": {
                "idTeam": {
                    "type": "integer",
                    "format": "int64"
                },
                "idUsu": {
                    "type": "integer",
                    "format": "int64"
                },
                "teamName": {
                    "type": "string"
                },
                "member_1": {
                    "type": "integer",
                    "format": "int64"
                },
                "member_2": {
                    "type": "integer",
                    "format": "int64"
                },
                "member_3": {
                    "type": "integer",
                    "format": "int64"
                },
                "member_4": {
                    "type": "integer",
                    "format": "int64"
                },
                "member_5": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
...
    }
}